Added methods to parse Cabbage Snaps;
Added CreatePreset method, to create a CsoundUnityPreset from a list of CsoundChannelControllers.
Added WritePreset method, to write a CsoundUnityPreset at a path.
Some cleaning on the Presets methods;
Now the presets from the editor can be set also if the number of channels have changed.
Also some formatting on the regions.